Pick a direction for the cucumber-jvm tests
