+++ console-runner-features/src/test/resources/info/javaspec/console/javaspec_in_process.feature
@@ -1,4 +1,4 @@
-Feature: Console Runner
+Feature: JavaSpec (in-process)
   As a developer who is working on some code that is covered by specs
   In order to know which specs are running, passing, and failing as well as how they are failing
   I want to run specs on the command line and see a text report of specs and their results
@@ -7,7 +7,8 @@ Feature: Console Runner
   In order to observe side effects such as which specs have been run
   I want to run some acceptance tests from the same JVM process as JavaSpec
 
-  Scenario: A ConsoleRunner should run specs and tell you what happened
+
+  Scenario: JavaSpec should run specs and tell you what happened
     Given I have a JavaSpec runner for the console
     And I have a Java class that defines a suite of lambda specs
     When I run the specs in that class
@@ -16,14 +17,14 @@ Feature: Console Runner
     And The runner should indicate which specs passed and failed
 
 
-  Scenario: A ConsoleRunner should report failing specs with its exit code
+  Scenario: JavaSpec should report failing specs with its exit code
     Given I have a JavaSpec runner for the console
     And I have a Java class that defines a suite of 1 or more failing lambda specs
     When I run the specs in that class
     Then The runner should indicate that 1 or more specs have failed
 
 
-  Scenario: A ConsoleRunner should report all passing specs with its exit code
+  Scenario: JavaSpec should report all passing specs with its exit code
     Given I have a JavaSpec runner for the console
     And I have a Java class that defines a suite of passing lambda specs
     When I run the specs in that class
