+++ features/command_line.feature
@@ -1,5 +1,5 @@
 @log_commands
-Feature: Command Line
+Feature: JavaSpec CLI (external process)
   As a developer who is working on some code that is covered by specs
   In order to know which specs are running, passing, and failing as well as how they are failing
   I want to run specs on the command line and see a text report of specs and their results
